pyproject.toml
[tool.poetry]
name = "acl-miniconf"
version = "0.1.0"
description = ""
authors = ["EntilZha <me@pedro.ai>"]
readme = "README.md"
packages = [{include = "acl_miniconf"}]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
Frozen-Flask = "^0.18"
Flask-Markdown = "^0.3"
rich = "^13.0.1"
PyYAML = "^6.0"
rocketchat-API = "^1.28.1"
transformers = "^4.25.1"
sentence-transformers = "^2.2.2"
scikit-learn = "^1.2.0"
torch = "^1.13.1"
ics = "^0.7.2"
hydra-core = "^1.3.1"


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
pylint = "^2.15.10"
mypy = "^0.991"
isort = "^5.11.4"
types-PyYAML = "^6.0.12.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
known_first_party = "pedroai"
profile = "black"
multi_line_output = 3